Upgrade to v0.12
One of the hacks used in this module does not translate to 0.12 very
well. This module is only used in one place in our config and, by
itself, does almost nothing. I'm fixing this for now to work only with
the one config that uses it, but we should delete this module.